all of the 3-6 color graphs are probably wrong

+++ main.py
@@ -1,5 +1,4 @@
 import numpy as np
-import scipy as sp
 import graphviz
 import os
 from tqdm import tqdm
@@ -31,13 +30,15 @@ def get_flipped_nodes(state: np.ndarray[int]) -> str:
 	return readout
 
 def compute_valid_states(edge_matrix: np.ndarray[int, int], num_colors: int = 2) -> np.ndarray[int]:
-	for num in tqdm(range(num_colors ** 10)):
+	numbers = list(range(num_colors ** 10))
+	numbers.sort(key=lambda x: bin(x).count('1'))
+	for num in tqdm(numbers):
 		move_array = np.array([0] * 10)
 		for idx in range(10):
 			move_array[idx] = num % num_colors
 			num >>= 1
 		state = edge_matrix.dot(move_array) % num_colors
-		yield state
+		yield state, num
 
 
 def get_node_position(index: int) -> str:
@@ -51,7 +52,7 @@ def get_node_position(index: int) -> str:
 
 
 def create_graph(edge_matrix: np.ndarray[np.ndarray[int]], state: np.ndarray[int], name: str = 'Petersen Graph', dir: str = 'graphs'):
-	colors = ['blue', 'red', 'green']
+	colors = ['blue', 'red', 'green', 'purple', 'aqua', 'goldenrod2']
 	graph = graphviz.Graph(name, directory=dir, filename=f'{name}.dot', engine='neato', format='png')
 	for i, vertex in enumerate(state):
 		graph.node(f'{i}', f'{i}', color = colors[vertex], pos = get_node_position(i))
@@ -64,6 +65,8 @@ def create_graph(edge_matrix: np.ndarray[np.ndarray[int]], state: np.ndarray[int
 
 def clean_graph_directory(dir: str = 'graphs'):
 	dir_name = f'C:\\Users\\Alex Prichard\\VSCodeProjects\\Python\\allBlue\\{dir}\\'
+	if not os.path.isdir(dir_name):
+		return
 	directory = os.listdir(dir_name)
 	for file in directory:
 		if file.endswith('.dot'):
@@ -72,6 +75,8 @@ def clean_graph_directory(dir: str = 'graphs'):
 
 def clear_graph_directory(dir: str = 'graphs'):
 	dir_name = f'C:\\Users\\Alex Prichard\\VSCodeProjects\\Python\\allBlue\\{dir}\\'
+	if not os.path.isdir(dir_name):
+		return
 	directory = os.listdir(dir_name)
 	for file in directory:
 		os.remove(os.path.join(dir_name, file))
@@ -109,6 +114,22 @@ def s4(state: np.ndarray[int]) -> np.ndarray[int]:
 	return permutation.dot(state)
 
 
+def s3(state: np.ndarray[int]) -> np.ndarray[int]:
+	permutation = np.array([
+		[0, 0, 0, 0, 0, 0, 0, 1, 0, 0],
+		[0, 0, 0, 0, 0, 0, 0, 0, 0, 1],
+		[0, 0, 0, 0, 0, 0, 1, 0, 0, 0],
+		[0, 1, 0, 0, 0, 0, 0, 0, 0, 0],
+		[0, 0, 1, 0, 0, 0, 0, 0, 0, 0],
+		[0, 0, 0, 0, 0, 1, 0, 0, 0, 0],
+		[0, 0, 0, 0, 1, 0, 0, 0, 0, 0],
+		[0, 0, 0, 0, 0, 0, 0, 0, 1, 0],
+		[1, 0, 0, 0, 0, 0, 0, 0, 0, 0],
+		[0, 0, 0, 1, 0, 0, 0, 0, 0, 0]
+	])
+	return permutation.dot(state)
+
+
 def hash_single_state(state: np.ndarray[int], num_colors: int = 2) -> int:
 	acc = 0
 	for node in state:
@@ -117,20 +138,29 @@ def hash_single_state(state: np.ndarray[int], num_colors: int = 2) -> int:
 	return acc
 
 
-def hash_state(state: np.ndarray[int]) -> int:
-	hashes = []
-	four_symmetry = state.copy()
-	for i in range(4):
-		four_symmetry = s4(four_symmetry)
-		five_symmetry = four_symmetry.copy()
-		for j in range(5):
-			five_symmetry = s5(five_symmetry)
-			hashes.append(hash_single_state(five_symmetry))
-	return min(hashes)
+def hash_state(state: np.ndarray[int], searched_graphs: set[int]) -> int:
+	min_hash = hash_single_state(state)
+	best_state = state
+	three_symmetry = state
+	for i in range(3):
+		three_symmetry = s3(three_symmetry)
+		four_symmetry = three_symmetry
+		for i in range(4):
+			four_symmetry = s4(four_symmetry)
+			five_symmetry = four_symmetry
+			for j in range(5):
+				five_symmetry = s5(five_symmetry)
+				current_hash = hash_single_state(five_symmetry)
+				searched_graphs.add(current_hash)
+				if current_hash < min_hash:
+					min_hash = current_hash
+					best_state = five_symmetry
+	return min_hash, best_state
 
 
 def main():
-	num_colors = 2
+	# 6, 120, 175, 205, 259
+	num_colors = 3
 	dir = f'{num_colors}_graphs'
 	edge_matrix = []
 	with open('edgeMatrix.txt') as file:
@@ -138,18 +168,23 @@ def main():
 		edge_matrix = np.array([[int(elem) for elem in line.split(' ')] for line in edge_matrix.split('\n')])
 
 	states = {}
-	for state in compute_valid_states(edge_matrix, num_colors):
-		state_hash = hash_state(state)
-		if state_hash in states:
+	searched_graphs = set()
+	max_moves = 0
+	for state, move in compute_valid_states(edge_matrix, num_colors):
+		if hash_single_state(state) in searched_graphs:
 			continue
-		states |= {state_hash: state.copy()}
+		state_hash, best_state = hash_state(state, searched_graphs)
+		states |= {state_hash: best_state}
+		max_moves = max(max_moves, bin(move).count('1'))
 	
 	print(f'Found {len(states)} states.')
 	state_values = list(states.values())
+	state_values.sort(key=lambda x: hash_single_state(x, num_colors))
 	clear_graph_directory(dir)
-	for i, state in tqdm(enumerate(state_values)):
+	for i, state in enumerate(tqdm(state_values)):
 		create_graph(edge_matrix, state, f'graph_{i}', dir)
 	clean_graph_directory(dir)
+	print(max_moves)
 
 if __name__ == '__main__':
 	main()
